Separate datafusion_poc into another crate
