+++ datafusion-integration/README.md
@@ -1,0 +1,1 @@
+Rules and examples to show how to integrate dolomite with datafusion planner.
