+++ dolomite/Cargo.toml
@@ -1,0 +1,32 @@
+[package]
+name = "dolomite"
+version = "0.1.0"
+edition = "2021"
+
+[dependencies]
+datafusion = { git = "https://github.com/apache/arrow-datafusion", rev = "884101e" }
+datafusion-physical-expr = { git = "https://github.com/apache/arrow-datafusion", rev = "884101e" }
+thiserror = "1"
+enum_dispatch = "0.3"
+log = "^0.4"
+enumset = "1"
+itertools = "0.10"
+enum-as-inner = "0.4"
+derive_more = "0.99"
+petgraph = "0.6"
+anyhow = "1"
+smallvec = "1"
+prettytable-rs = "^0.8"
+lazy_static = "1"
+strum = "0.24"
+strum_macros = "0.24"
+async-trait = "0.1.41"
+futures = "0.3"
+
+
+[dev-dependencies]
+serde = "1.0"
+serde_json = "1.0"
+tokio = { version = "1.0", features = ["macros", "rt", "rt-multi-thread", "sync", "fs", "parking_lot"] }
+tokio-stream = "0.1"
+
