+++ dolomite/src/rules/pattern.rs
