+++ dolomite/src/rules/table_scan.rs
