+++ dolomite/src/stat.rs
